generics revision

+++ task02array/src/test/java/jwd/task02array/ArrayCreatorTest.java
@@ -15,18 +15,18 @@ public class ArrayCreatorTest {
 
 	@DataProvider(name = "DataForArrayCreating")
 	public Object[][] createDataForArrayCreating() {
-		return new Object[][] { { "dataTest1.txt", new Integer[] { 1 } },
-				{ "dataTest2.txt", new Integer[] { 0, 0, 0, 0, 0 } },
-				{ "dataTest3.txt", new Integer[] { 10, -8, 0, 10, 7 } }, { "dataTest4.txt", new Integer[] { 1, 2 } } };
+		return new Object[][] { { "dataTest1.txt", new Double[] { 1.0 } },
+				{ "dataTest2.txt", new Double[] { 0.0, 0.0, 0.0, 0.0, 0.0 } },
+				{ "dataTest3.txt", new Double[] { 10.0, -8.0, 0.0, 10.0, 7.0 } }, { "dataTest4.txt", new Double[] { 1.0, 2.0 } } };
 	}
 
 	@Test(groups = { "service" }, dataProvider = "DataForArrayCreating")
-	public void testDataForArrayCreating(String a, Integer[] b) throws ServiceException {
+	public void testDataForArrayCreating(String a, Double[] b) throws ServiceException {
 		String fileName = a;
-		Array<Integer> array = new Array<>(b);
+		Array<Double> array = new Array<>(b);
 
-		Array<Integer> actual = service.createArrayFromFile(fileName);
-		Array<Integer> expected = array;
+		Array<Double> actual = service.createArrayFromFile(fileName);
+		Array<Double> expected = array;
 		assertEquals(actual, expected);
 	}
 }
